added CORS support. Now front end is connected to back end.

+++ BankController.mjs
@@ -3,7 +3,10 @@ import BankService from "./BankService.mjs";
 
 class BankController {
     async create(req, res) {
+        res.set('Access-Control-Allow-Origin', '*');
+
         console.log('create')
+        console.log(req.body)
         try {
             const bank = await BankService.create(req.body)
             console.log(bank)
@@ -14,6 +17,7 @@ class BankController {
     }
 
     async getAll(req, res) {
+        res.set('Access-Control-Allow-Origin', '*');
         try {
             const banks = await BankService.getAll();
             return res.json(banks);
@@ -23,6 +27,8 @@ class BankController {
     }
 
     async getOne(req, res) {
+        res.set('Access-Control-Allow-Origin', '*');
+
         try {
             const bank = await BankService.getOne(req.params.id);
             return res.json(bank)
@@ -32,6 +38,8 @@ class BankController {
     }
 
     async update(req, res) {
+        res.set('Access-Control-Allow-Origin', '*');
+        console.log(req.body);
         try {
             const bank = req.body;
             if (!bank._id)
@@ -44,10 +52,16 @@ class BankController {
     }
 
     async delete(req, res) {
+        res.set('Access-Control-Allow-Origin', '*');
+
         try {
             const id = req.params.id;
             if (!id)
                 return res.status(400).json({message: "Missing id"});
+
+            if(id ==="625ef01d6e187a99202d531b")
+                return res.status(400).json({message: "You can't delete the default bank"});
+
             const deletedBank = await BankService.delete(id);
             return res.json(deletedBank);
